Refactor: move sources to src and fix listUser typo
