--- Data.js
@@ -1,6 +1,0 @@
-export class Data {
-  constructor(id, taille) {
-    this.id = id; //Un entier unique
-    this.taille = taille; //Un entier
-  }
-}
