--- Node.js
@@ -1,5 +1,0 @@
-export class Node {
-    constructor(id) {
-      this.id = id; // Un entier unique
-    }
-  }
