+++ package.json
@@ -2,10 +2,11 @@
     "name": "stockage-donnes-personelles",
     "version": "1.0.0",
     "description": "Permet de représenter le stockage de données personnelles dans une structure de données",
-    "main": "main.js",
+    "main": "src/main.js",
     "type": "module",
     "scripts": {
-        "start": "node main.js"
+        "start": "node src/main.js",
+        "test": "node --test"
     },
     "author": "Arthur Rattanavong",
     "license": "ISC"
