+++ src/Node.js
@@ -1,0 +1,8 @@
+/**
+ * Représente un nœud générique identifié uniquement par un id.
+ */
+export class Node {
+  constructor(id) {
+    this.id = id; // Un entier unique
+  }
+}
